+++ lib/fp_util.dart
@@ -7,6 +7,7 @@ export 'src/extensions/extensions.dart';
 export 'src/form/form.dart';
 export 'src/helpers/clipboard_helper.dart';
 export 'src/helpers/image_picker_helper.dart';
+export 'src/helpers/uuid_helper.dart';
 export 'src/input_formatters/input_formatters.dart';
 export 'src/logger/logger.dart';
 export 'src/responsive/responsive.dart';